refactor: rewrite ads and analytics endpoints to the new system
